RDA5815 + FFT

+++ tlv320aic3204.c
@@ -64,7 +64,7 @@ static const uint8_t conf_data_routing[] = {
   2, 0x01, 0x08, /* Disable Internal Crude AVdd in presence of external AVdd supply or before powering up internal AVdd LDO*/
   2, 0x02, 0x01, /* Enable Master Analog Power Control */
   2, 0x7b, 0x01, /* Set the REF charging time to 40ms */
-  2, 0x14, 0x25, /* HP soft stepping settings for optimal pop performance at power up Rpop used is 6k with N = 6 and soft step = 20usec. This should work with 47uF coupling capacitor. Can try N=5,6 or 7 time constants as well. Trade-off delay vs “pop” sound. */
+  2, 0x14, 0x25, /* HP soft stepping settings for optimal pop performance at power up Rpop used is 6k with N = 6 and soft step = 20usec. This should work with 47uF coupling capacitor. Can try N=5,6 or 7 time constants as well. Trade-off delay vs sound. */
   2, 0x0a, 0x33, /* Set the Input Common Mode to 0.9V and Output Common Mode for Headphone to 1.65V */
 
   2, 0x3d, 0x00, /* Select ADC PTM_R4 */
@@ -125,7 +125,7 @@ void tlv320aic3204_init(void)
   tlv320aic3204_config(conf_data_unmute);
 }
 
-void tlv320aic3204_select(int channel)
+void tlv320aic3204_select(uint8_t channel)
 {
     const uint8_t ch3[] = {
         2, 0x00, 0x01, /* Select Page 1 */
@@ -142,7 +142,7 @@ void tlv320aic3204_select(int channel)
     tlv320aic3204_config(channel ? ch1 : ch3);
 }
 
-void tlv320aic3204_set_gain(int lgain, int rgain)
+void tlv320aic3204_set_gain(uint8_t lgain, uint8_t rgain)
 {
     uint8_t data[] = {
         2, 0x00, 0x01, /* Select Page 1 */
